+++ src/Move/StartOfDay.php
@@ -5,10 +5,7 @@ namespace Innmind\TimeContinuum\Move;
 
 use Innmind\TimeContinuum\{
     PointInTime,
-    Period\Hour,
-    Period\Minute,
-    Period\Second,
-    Period\Millisecond,
+    Period,
 };
 
 final class StartOfDay
@@ -16,10 +13,10 @@ final class StartOfDay
     public function __invoke(PointInTime $point): PointInTime
     {
         return $point->goBack(
-            Hour::of($point->hour()->toInt())
-                ->add(Minute::of($point->minute()->toInt()))
-                ->add(Second::of($point->second()->toInt()))
-                ->add(Millisecond::of($point->millisecond()->toInt())),
+            Period::hour($point->hour()->toInt())
+                ->add(Period::minute($point->minute()->toInt()))
+                ->add(Period::second($point->second()->toInt()))
+                ->add(Period::millisecond($point->millisecond()->toInt())),
         );
     }
 }